fix(mcp): move CallTool to test-only file, acknowledge Close error
